requirements.txt
python-dateutil~=2.9.0
python-dotenv~=1.0.0
Flask-Login~=0.6.3
Flask~=3.1.2
Flask-CORS~=4.0.0
gunicorn~=23.0.0
Flask-SQLAlchemy~=3.1.1
Werkzeug~=3.1.4
SQLAlchemy~=2.0.45
psycopg2-binary~=2.9.11
itsdangerous~=2.2.0
pytz~=2025.2
flask-migrate~=4.1.0
flask-talisman~=1.1.0
WTForms~=3.2.1
alembic~=1.16.5
Flask-WTF~=1.2.2
markupsafe~=3.0.3
requests~=2.32.5
email-validator~=2.3.0